course_mcp/3. MCP with LangChain/requirements.txt
langchain==1.2.13 
langchain-community==0.4.1
langchain-openai==1.1.11 
langchain-mcp-adapters==0.2.2 
elasticsearch==8.13.0
python-dotenv==1.1.0
fastmcp==3.1.1
mcp==1.26.0
